# Add "normalize" method to Group and improve performance of serialization