--- a/README.md
+++ b/README.md
@@ -14,7 +14,7 @@ Add the `bn` crate to your dependencies in `Cargo.toml`...
 
 ```toml
 [dependencies]
-bn = "0.4.0"
+bn = "0.4.1"
 ```
 
 ...and add an `extern crate` declaration to your crate root: