This adds a new `normalize` method to `Group` that can be used to speed up serialization in parallel. The jacobian-to-affine conversion will handle the trivial case of z=1, speeding up serialization significantly in some cases. In the future, I would rather elevate this stuff to the type system, but this works for now.

Also, cleaned up the benchmarks.

```
test fq12_exponentiation         ... bench:   5,969,704 ns/iter (+/- 410,626)
test fq12_scalar_multiplication  ... bench:      19,286 ns/iter (+/- 756)
test fr_addition                 ... bench:          22 ns/iter (+/- 5)
test fr_inverses                 ... bench:      10,260 ns/iter (+/- 206)
test fr_multiplication           ... bench:         146 ns/iter (+/- 47)
test fr_subtraction              ... bench:          22 ns/iter (+/- 6)
test g1_addition                 ... bench:       2,597 ns/iter (+/- 200)
test g1_deserialization          ... bench:       1,229 ns/iter (+/- 118)
test g1_scalar_multiplication    ... bench:     633,965 ns/iter (+/- 12,880)
test g1_serialization            ... bench:      11,888 ns/iter (+/- 261)
test g1_serialization_normalized ... bench:         687 ns/iter (+/- 169)
test g1_subtraction              ... bench:       2,606 ns/iter (+/- 152)
test g2_addition                 ... bench:      11,340 ns/iter (+/- 475)
test g2_deserialization          ... bench:       9,407 ns/iter (+/- 273)
test g2_scalar_multiplication    ... bench:   2,728,459 ns/iter (+/- 336,310)
test g2_serialization            ... bench:      15,467 ns/iter (+/- 257)
test g2_serialization_normalized ... bench:       1,353 ns/iter (+/- 146)
test g2_subtraction              ... bench:      11,382 ns/iter (+/- 238)
test perform_pairing             ... bench:   6,992,153 ns/iter (+/- 788,769)
```